+++ src/engine/scene/Sky.cpp
@@ -8,7 +8,7 @@ Sky::Sky(const Storage mode) : Object(Object::Type::Common, Resources::manager()
 
 void Sky::decode(const std::vector<KeyValues> & params, const Storage mode){
 	Object::decode(params, mode);
-	for(int pid = 0; pid < params.size(); ++pid){
+	for(size_t pid = 0; pid < params.size(); ++pid){
 		const auto & param = params[pid];
 		if(param.key == "direction"){
 			_sunDirection = Codable::decodeVec3(param);
